requirements.txt
python-telegram-bot==22.1
nest_asyncio
python-dotenv
requests
beautifulsoup4
schedule
